task 1 v3

+++ python_async_function/1-concurrent_coroutines.py
@@ -8,7 +8,7 @@ async def wait_random(max_delay=10):
     await asyncio.sleep(delay)
     return delay
 
-async def wait_n(n: int, max_delay: float) -> List[float]:
+async def wait_n(n: int, max_delay: int) -> List[float]:
     tasks = [wait_random(max_delay) for _ in range(n)]
     results = await asyncio.gather(*tasks)
     return sorted(results)
